requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Django==1.11.7
django-cors-headers==2.1.0
djangorestframework==3.7.3
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2017.3
